Refactor app to use OOP constructs
