+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = git-trend
-version = 0.5.1
+version = 0.5.2
 author = Manoj Karthick
 author_email = manojkarthick@ymail.com
 url = https://github.com/manojkarthick/git-trend
@@ -21,10 +21,13 @@ install_requires =
     beautifulsoup4
     requests
     prettytable
-py_modules = get_trending_items
+py_modules =
+    trending
+    utils
+    enums
 python_requires = >=3.6
 
 
 [options.entry_points]
 console_scripts =
-    git-trend = get_trending_items:cli
+    git-trend = trending:cli
